pages build fixed.
