+++ package.json
@@ -22,6 +22,7 @@
     "@types/mdx": "^2.0.12",
     "contentlayer": "^0.3.4",
     "framer-motion": "^11.0.8",
+    "fs": "^0.0.1-security",
     "gray-matter": "^4.0.3",
     "locomotive-scroll": "^5.0.0-beta.11",
     "markdown-to-jsx": "^7.4.7",
